+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## v0.95dev: UNDER DEVELOPMENT
 __No Release Date Set__
 
-* 
+* Copy `global/pg_control` last during backups.
+
+* Write `.info` and `.manifest` files to temp before moving them to their final locations and fsync'ing.
 
 ## v0.90: 9.5 Support, Various Enhancements, and Minor Bug Fixes
 __Released February 7, 2016__